pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "psycopg2-binary>=2.9.10",
    "sqlite-web>=0.6.4",
]
